requirements.txt
imapclient==2.3.1
spacy==3.7.4
pandas==2.0.3
python-dotenv==1.0.0
geopy==2.4.1
python-dateutil==2.8.2